requirements.txt
pytest
pytest-html
requests
pycryptodome
